store: move common functions to utils
